requirements.txt
torch
fastapi
uvicorn
pillow
torchvision
scikit-learn
pandas
tqdm
python-multipart
yaml
pytest
